Refactoring to use components and pages
