+++ src/App.js
@@ -1,8 +1,8 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import GlobalStyles from "./assets/styles/GlobalStyles";
-import Login from "./components/Login/Login";
-import SignUp from "./components/SignUp/SignUp";
-import Home from "./components/Home/Home";
+import Login from "./pages/Login/Login";
+import SignUp from "./pages/SignUp/SignUp";
+import Home from "./pages/Home/Home";
 import PrivatePage from "./components/privatePage/PrivatePage";
 
 export default function App() {
